Move print function in a separate file in order to be shared with linux_platform
